Removed VS2010 project, fixed #ifdef for CLI
So now I'm working on breaking the SQL string generation out of the
existing methods for the purpose of allowing developers to use either
the sqlite3 c files (which are incompatible with CLI) or the sqlite DLL
(which is compatible with CLI). Basically this means developers will be
able to make GUI applications using Visual Studio or something else.